requirements.txt
AC-Flask-HipChat==0.2.2
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.2
pymongo==3.3.0
redis==2.10.5
requests==2.11.1
Werkzeug==0.11.10